+++ mutwo/core_parameters/__init__.py
@@ -2,6 +2,7 @@
 
 """
 
+from . import constants
 from . import configurations
 from . import abc