+++ mutwo/core_parameters/constants.py
@@ -1,0 +1,9 @@
+import typing
+
+TempoInBeatsPerMinute = float
+TempoRangeInBeatsPerMinute = tuple[TempoInBeatsPerMinute, TempoInBeatsPerMinute]
+TempoOrTempoRangeInBeatsPerMinute = typing.Union[
+    TempoInBeatsPerMinute, TempoRangeInBeatsPerMinute
+]
+
+del typing